Add new com.student.base Package and TestBase.java class
