+++ src/test/java/com/student/test/StudentDeleteTest.java
@@ -1,28 +1,15 @@
 package com.student.test;
 
-import static org.hamcrest.Matchers.*;
 import static com.jayway.restassured.RestAssured.*;
 
-import java.util.ArrayList;
-
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.jayway.restassured.RestAssured;
-import com.jayway.restassured.http.ContentType;
-import com.student.model.Student;
 
-public class StudentDeleteTest {
+import com.student.base.TestBase;
 
-	@BeforeClass
-	public static void init() {
 
-		RestAssured.baseURI = "http://localhost";
-		RestAssured.port = 8080;
-		RestAssured.basePath = "/student";
-	
-	}
-	
+public class StudentDeleteTest extends TestBase {
+
 	@Test
 	public void deleteStudent() {
 		given()
