+++ src/test/java/com/student/test/StudentPostTest.java
@@ -1,28 +1,19 @@
 package com.student.test;
 
-import org.junit.BeforeClass;
+import static com.jayway.restassured.RestAssured.*;
+
 import org.junit.Test;
 
-import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.http.ContentType;
+import com.student.base.TestBase;
 import com.student.model.Student;
-
-
 
 import java.util.ArrayList;
 
-import static com.jayway.restassured.RestAssured.*;
 
-public class StudentPostTest {
 
-	@BeforeClass
-	public static void init() {
+public class StudentPostTest extends TestBase {
 
-		RestAssured.baseURI = "http://localhost";
-		RestAssured.port = 8080;
-		RestAssured.basePath = "/student";
-	}
-	
 	/**
 	 * 
 	 */
