+++ src/test/java/com/student/test/StudentPutTest.java
@@ -1,27 +1,16 @@
 package com.student.test;
 
-import static org.hamcrest.Matchers.*;
 import static com.jayway.restassured.RestAssured.*;
 
 import java.util.ArrayList;
 
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.http.ContentType;
+import com.student.base.TestBase;
 import com.student.model.Student;
 
-public class StudentPutTest {
-
-	@BeforeClass
-	public static void init() {
-
-		RestAssured.baseURI = "http://localhost";
-		RestAssured.port = 8080;
-		RestAssured.basePath = "/student";
-	
-	}
+public class StudentPutTest extends TestBase{
 	
 	
 	@Test
